+++ sample/src/main/java/me/denley/preferenceinjector/sample/SampleFragment.java
@@ -1,4 +1,4 @@
-package me.denley.preferenceinjector;
+package me.denley.preferenceinjector.sample;
 
 import android.app.Fragment;
 import android.content.SharedPreferences;
@@ -13,6 +13,12 @@ import android.widget.CheckBox;
 import android.widget.SeekBar;
 
 import java.util.Set;
+
+import me.denley.preferenceinjector.InjectPreference;
+import me.denley.preferenceinjector.OnPreferenceChange;
+import me.denley.preferenceinjector.PreferenceDefault;
+import me.denley.preferenceinjector.PreferenceInjector;
+import me.denley.preferenceinjector.R;
 
 /**
  * Created by Denley on 25/02/2015.
@@ -52,7 +58,7 @@ public class SampleFragment extends Fragment {
         final View view = inflater.inflate(R.layout.activity_main, container, false);
         booleanPreferenceDisplay = (CheckBox) view.findViewById(R.id.pref_boolean);
         integerPreferenceDisplay = (SeekBar) view.findViewById(R.id.pref_integer);
-        PreferenceInjector.inject(getActivity(), this);
+        PreferenceInjector.inject(this);
 
         startHandlerOnBackgroundThread();
         return view;