+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+### Added
+- `PreferenceInjector.inject(Fragment target)` method. The `Context` can be inferred from the `Fragment`.
+
 ## 2.1.0 - 2015-02-19
 ### Added
 - `@PreferenceDefault` annotation, to specify injection values when none exists in the `SharedPreferences` file